lazy standard module checks
